rename commandPopover -> shortcutSnippetMenu
