+++ src/contentScript/shortcutSnippetMenu/CommandPopover.scss
