+++ src/contentScript/shortcutSnippetMenu/CommandRegistry.ts
