+++ src/contentScript/shortcutSnippetMenu/commandUtils.ts
