+++ src/contentScript/shortcutSnippetMenu/useKeyboardQuery.ts
