+++ src/tsconfig.strictNullChecks.json
@@ -296,17 +296,17 @@
     "./components/walkthroughModal/showWalkthroughModal.ts",
     "./contentScript/activationConstants.ts",
     "./contentScript/browserActionInstantHandler.ts",
-    "./contentScript/commandPopover/CommandPopover.stories.tsx",
-    "./contentScript/commandPopover/CommandPopover.test.tsx",
-    "./contentScript/commandPopover/CommandPopover.tsx",
-    "./contentScript/commandPopover/CommandRegistry.ts",
-    "./contentScript/commandPopover/commandController.tsx",
-    "./contentScript/commandPopover/commandPopoverSlice.ts",
-    "./contentScript/commandPopover/commandPopoverSlice.test.ts",
-    "./contentScript/commandPopover/commandUtils.ts",
-    "./contentScript/commandPopover/commandUtils.test.ts",
-    "./contentScript/commandPopover/useCommandRegistry.ts",
-    "./contentScript/commandPopover/useKeyboardQuery.ts",
+    "./contentScript/shortcutSnippetMenu/CommandPopover.stories.tsx",
+    "./contentScript/shortcutSnippetMenu/CommandPopover.test.tsx",
+    "./contentScript/shortcutSnippetMenu/CommandPopover.tsx",
+    "./contentScript/shortcutSnippetMenu/CommandRegistry.ts",
+    "./contentScript/shortcutSnippetMenu/commandController.tsx",
+    "./contentScript/shortcutSnippetMenu/commandPopoverSlice.ts",
+    "./contentScript/shortcutSnippetMenu/commandPopoverSlice.test.ts",
+    "./contentScript/shortcutSnippetMenu/commandUtils.ts",
+    "./contentScript/shortcutSnippetMenu/commandUtils.test.ts",
+    "./contentScript/shortcutSnippetMenu/useCommandRegistry.ts",
+    "./contentScript/shortcutSnippetMenu/useKeyboardQuery.ts",
     "./contentScript/context.ts",
     "./contentScript/contextMenus.ts",
     "./contentScript/elementReference.ts",
